Finished collections screen Tweaked ui Fixing some db/file locations on desktop

+++ data/network/public/src/commonMain/kotlin/app.campfire.network/models/FileMetadata.kt
@@ -21,7 +21,7 @@ data class FileMetadata(
   val path: String,
   val relPath: String,
   val size: Int,
-  val mtimeMs: Int,
-  val ctimeMs: Int,
-  val birthtimeMs: Int,
+  val mtimeMs: Long,
+  val ctimeMs: Long,
+  val birthtimeMs: Long,
 )
